third_party/WebKit/LayoutTests/fast/files/blob-constructor.html

<!DOCTYPE html>

<script src="../../resources/js-test.js"></script>
<script>
description("Test the Blob constructor.");
var jsTestIsAsync = true;

// Test that the File-specific lastModified is not set by the Blob constructor.
shouldBe("(new Blob([])).lastModified", "undefined");
shouldBe("(new Blob([], {})).lastModified", "undefined");
shouldBe("(new Blob([], {lastModified: new Date()})).lastModified", "undefined");

// Test ArrayBufferView parameters.
shouldBe("new Blob([new DataView(new ArrayBuffer(100))]).size", "100");
shouldBe("new Blob([new Uint8Array(100)]).size", "100");
shouldBe("new Blob([new Uint8ClampedArray(100)]).size", "100");
shouldBe("new Blob([new Uint16Array(100)]).size", "200");
shouldBe("new Blob([new Uint32Array(100)]).size", "400");
shouldBe("new Blob([new Int8Array(100)]).size", "100");
shouldBe("new Blob([new Int16Array(100)]).size", "200");
shouldBe("new Blob([new Int32Array(100)]).size", "400");
shouldBe("new Blob([new Float32Array(100)]).size", "400");
shouldBe("new Blob([new Float64Array(100)]).size", "800");
shouldBe("new Blob([new Float64Array(100), new Int32Array(100), new Uint8Array(100), new DataView(new ArrayBuffer(100))]).size", "1400");
shouldBe("new Blob([new Blob([new Int32Array(100)]), new Uint8Array(100), new Float32Array(100), new DataView(new ArrayBuffer(100))]).size", "1000");

// Test ArrayBuffer parameters.
shouldBe("new Blob([(new DataView(new ArrayBuffer(100))).buffer]).size", "100");
shouldBe("new Blob([(new Uint8Array(100)).buffer]).size", "100");
shouldBe("new Blob([(new Uint8ClampedArray(100)).buffer]).size", "100");
shouldBe("new Blob([(new Uint16Array(100)).buffer]).size", "200");
shouldBe("new Blob([(new Uint32Array(100)).buffer]).size", "400");
shouldBe("new Blob([(new Int8Array(100)).buffer]).size", "100");
shouldBe("new Blob([(new Int16Array(100)).buffer]).size", "200");
shouldBe("new Blob([(new Int32Array(100)).buffer]).size", "400");
shouldBe("new Blob([(new Float32Array(100)).buffer]).size", "400");
shouldBe("new Blob([(new Float64Array(100)).buffer]).size", "800");
shouldBe("new Blob([(new Float64Array(100)).buffer, (new Int32Array(100)).buffer, (new Uint8Array(100)).buffer, (new DataView(new ArrayBuffer(100))).buffer]).size", "1400");
shouldBe("new Blob([new Blob([(new Int32Array(100)).buffer]), (new Uint8Array(100)).buffer, (new Float32Array(100)).buffer, (new DataView(new ArrayBuffer(100))).buffer]).size", "1000");

if (window.SharedArrayBuffer) {
  // Test SharedArrayBuffer parameters.
  shouldThrow("new Blob([new Uint8Array(new SharedArrayBuffer(4))])", '"TypeError: Failed to construct \'Blob\': The provided ArrayBufferView value must not be shared."');
}

testNormalization();

function testNormalization() {
  // Test that strings are not NFC normalized
  OMICRON_WITH_OXIA = '\u1F79'; // NFC normalized to U+3CC
  shouldBe("OMICRON_WITH_OXIA.charCodeAt(0)", "0x1F79");
  reader = new FileReader();
  reader.readAsText(new Blob([OMICRON_WITH_OXIA]));
  reader.onload = function() {
    shouldBe("reader.result.charCodeAt(0)", "0x1F79");
    testEncodingReplacements();
  };
}

function testEncodingReplacements() {
  // Test that invalid UTF-16 code units are replaced.
  CONTAINS_UNPAIRED_SURROGATES = 'abc\uDC00def\uD800ghi';
  shouldBe("CONTAINS_UNPAIRED_SURROGATES.charCodeAt(3)", "0xDC00");
  shouldBe("CONTAINS_UNPAIRED_SURROGATES.charCodeAt(7)", "0xD800");
  reader = new FileReader();
  reader.readAsText(new Blob([CONTAINS_UNPAIRED_SURROGATES]));
  reader.onload = function() {
    shouldBe("reader.result.charCodeAt(3)", "0xFFFD");
    shouldBe("reader.result.charCodeAt(7)", "0xFFFD");
    finishJSTest();
  };
}

</script>
